+++ TabletopTweaks/NewContent/ContentAdder.cs
@@ -17,12 +17,18 @@ namespace TabletopTweaks.NewContent {
                 Features.MartialWeaponProficencySelection.AddMartialWeaponProficencySelection();
                 //Features.NauseatedPoision.AddNauseatedPoision();
 
+                Spells.LongArms.AddLongArms();
+                Spells.ShadowEnchantment.AddShadowEnchantment();
+                Spells.ShadowEnchantment.AddShadowEnchantmentGreater();
+                Spells.MagicalTailSpells.AddNewMagicalTailSpells();
+
                 //Added early as some things depend on them for selections
                 Feats.IntensifiedSpell.AddIntensifiedSpell();
                 Feats.RimeSpell.AddRimeSpell();
                 Feats.BurningSpell.AddBurningSpell();
                 Feats.FlaringSpell.AddFlaringSpell();
                 Feats.PiercingSpell.AddPiercingSpell();
+                Feats.SolidShadows.AddSolidShadows();
 
                 Templates.AlignmentTemplates.AddCelestialTemplate();
                 Templates.AlignmentTemplates.AddEntropicTemplate();
@@ -95,12 +101,6 @@ namespace TabletopTweaks.NewContent {
                 Archetypes.Myrmidarch.AddMyrmidarch();
                 //Features to support existing prestige clases
                 Classes.Loremaster.AddLoremasterFeatures();
-
-
-                Spells.LongArms.AddLongArms();
-                Spells.ShadowEnchantment.AddShadowEnchantment();
-                Spells.ShadowEnchantment.AddShadowEnchantmentGreater();
-                Spells.MagicalTailSpells.AddNewMagicalTailSpells();
 
                 MythicAbilities.ImpossibleSpeed.AddImpossibleSpeed();
                 MythicAbilities.ArmorMaster.AddArmorMaster();